Fix file name for ReplRuntime and move to runtime folder
